+++ dictdiffer.py
@@ -1,222 +1,175 @@
-import copy
-import collections
+from copy import deepcopy
 
-(ADD, REMOVE, PUSH, PULL, CHANGE) = (
-    'add', 'remove', 'push', 'pull', 'change')
+class DictDiffer(object):
+    """ Base Class for a DictDiffer
 
+    To add also diff functions for other types like strings or whatever, just
+    subclass ``DictDiffer`` and provide methodes with a name pattern like
+    "diff_<new implemented type.__name__>", then it will be called for this type.
 
-def diff (first, second, node=None):
+    Almost all the logic is taken from the dictdiffer_ module from Fatih
+    Erikli, but the implementation as a class has the advantage of making
+    custom diffs easy and even has better performance[#1]_
+
+    .. dictdiffer: https://github.com/fatiherikli/dictdiffer
+    .. [#1] run benchmarking.py if you don't believe me
+
     """
-    Compares two dictionary object, and returns a diff result.
+    ADD = 'add'
+    REMOVE = 'remove'
+    PUSH = 'push'
+    PULL = 'pull'
+    CHANGE = 'change'
+    METHODS = (ADD,REMOVE,PUSH,CHANGE,PULL)
+    SEP = '.'
+    def __init__(self):
+        self.patchers = {x:getattr(self,'_patch_%s'%x) for x in self.METHODS}
+        self.swappers = {x:getattr(self,'_swap_%s'%x) for x in self.METHODS}
 
-        >>> result = diff({'a':'b'}, {'a':'c'})
-        >>> list(result)
-        [('change', 'a', ('b', 'c'))]
+    def _join_nodes(self,node):
+        """ joins keys of subdictionaries """
+        return self.SEP.join(node)
 
-    """
-    node = node or []
-    dotted_node = '.'.join(node)
+    def _set_nodes(self,destination,lookup,value):
+        keys = lookup.split(self.SEP)
+        if len(keys) > 1:
+            dest = destination
+            for key in keys[:-1]:
+                dest = dest[key]
+            dest[keys[-1]] = value
+        else:
+            destination[keys[0]] = value
 
-    if isinstance(first, collections.Iterable):
-        # dictionaries are not hashable, we can't use sets
-        intersection = [k for k in first if k in second]
-        addition = [k for k in second if not k in first]
-        deletion = [k for k in first if not k in second]
+    def _lookup_nodes(self,source,lookup,parent=None):
+        """
+        A helper function that allows you to reach dictionary
+        items with dot lookup (e.g. document.properties.settings)
 
-    def  diff_dict():
-        """Compares if object is a dictionary. Callees again the parent
-        function as recursive if dictionary have child objects.
-        Yields `add` and `remove` flags."""
+            >>> dot_lookup({'a': {'b': 'hello'}}, 'a.b')
+            'hello'
+
+        If parent argument is True, returns the parent node of matched
+        object.
+
+            >>> dot_lookup({'a': {'b': 'hello'}}, 'a.b', parent=True)
+            {'b': 'hello'}
+
+        If node is empy value, returns the whole dictionary object.
+
+            >>> dot_lookup({'a': {'b': 'hello'}}, '')
+            {'a': {'b': 'hello'}}
+
+        """
+        if not lookup:
+            return source
+
+        value = source
+        keys = lookup.split('.')
+
+        if parent:
+            keys = keys[:-1]
+
+        for key in keys:
+            value = value[key]
+        return value
+
+    def diff_dict(self,first,second,node):
+        """ creates a diff of a dictionary """
+        # calculate difference in keys
+        firstset = frozenset(first)
+        secondset = frozenset(second)
+        intersection = firstset.intersection(secondset)
+        addition = secondset - intersection
+        deletion = firstset - intersection
+        # now produce the nodes
         if addition:
-            yield ADD, dotted_node, [
-                # for additions, return a list that consist with
-                # two-pair tuples.
-                (key, second[key]) for key in addition]
+            yield self.ADD,self._join_nodes(node),[(key,second[key]) for key in addition]
 
         if deletion:
-            yield REMOVE, dotted_node, [
-                # for deletions, return the list of removed keys
-                # and values.
-                (key, first[key]) for key in deletion]
+            yield self.REMOVE,self._join_nodes(node),[(key,first[key]) for key in deletion]
 
         for key in intersection:
-            # if type is not changed, callees again diff function to compare.
-            # otherwise, the change will be handled as `change` flag.
-            recurred = diff(
-                first[key],
-                second[key],
-                node=node + [key])
+            recurred = self.diff(first[key],second[key],node=node+[key])
+            for diff in recurred:
+                yield diff
 
-            for diffed in recurred:
-                yield diffed
+            # python 3
+            #yield from self.diff(first[key],second[key],node=node+[key])
 
-    def diff_list():
-        """Compares if objects are list. Yields `push` and `pull` flags."""
+    def diff_list(self,first,second,node):
+        """ creates a diff of a list """
+        addition = [k for k in second if not k in first]
+        deletion = [k for k in first if not k in second]
         if addition:
-            # the addition will be consist with the list of added values.
-            yield PUSH, dotted_node, addition
-
+            yield self.PUSH, self._join_nodes(node), addition
         if deletion:
-            # for deletions, returns the list of removed items
-            yield PULL, dotted_node, deletion
+            yield self.PULL, self._join_nodes(node), deletion
 
-    def diff_otherwise():
-        """Compares string and integer types. Yields `change` flag."""
+    def diff_generic(self,first,second,node):
+        """ creates a diff of all the other types not implemented """
         if first != second:
-            yield CHANGE, dotted_node, (first, second)
+            yield self.CHANGE, self._join_nodes(node), (first,second)
 
-    differs = {
-        dict: diff_dict,
-        list: diff_list
-    }
+    def diff(self,first,second,node=[]):
+        if type(first) == type(second):
+            return getattr(self,'diff_%s'%type(first).__name__,self.diff_generic)(first,second,node)
+        else:
+            return self.diff_generic(first,second,node)
 
-    differ = differs.get(type(first))
-    return (differ or diff_otherwise)()
-
-
-def patch(diff_result, destination):
-    """
-    Patches the diff result to the old dictionary.
-
-        >>> patch([('push', 'numbers', [2, 3])], {'numbers': []})
-        {'numbers': [2, 3]}
-
-        >>> patch([('pull', 'numbers', [1])], {'numbers': [1, 2, 3]})
-        {'numbers': [2, 3]}
-
-    """
-    destination = copy.deepcopy(destination)
-
-    def add(node, changes):
+    def _patch_add(self,destination,node,changes):
         for key, value in changes:
-            dot_lookup(destination, node)[key] = value
+            self._lookup_nodes(destination,node)[key] = value
 
-    def change(node, changes):
-        dest = dot_lookup(destination, node, parent=True)
-        last_node = node.split('.')[-1]
-        _, value = changes
-        dest[last_node] = value
+    def _patch_change(self,destination,node,changes):
+        dest = self._lookup_nodes(destination,node,parent=True)
+        try:
+            #dest[dest.keys()[0]] = changes[0]
+            self._set_nodes(destination,node,changes[1])
+        except Exception as e:
+            print(node,e)
 
-    def remove(node, changes):
-        for key, _ in changes:
-            del dot_lookup(destination, node)[key]
+    def _patch_remove(self,destination,node,changes):
+        for key,value in changes:
+            del self._lookup_nodes(destination,node)[key]
 
-    def pull(node, changes):
-        dest = dot_lookup(destination, node)
-        for val in changes:
-            dest.remove(val)
+    def _patch_pull(self,destination,node,changes):
+        dest = self._lookup_nodes(destination,node)
+        for value in changes:
+            dest.remove(value)
 
-    def push(node, changes):
-        dest = dot_lookup(destination, node)
-        for val in changes:
-            dest.append(val)
+    def _patch_push(self,destination,node,changes):
+        dest = self._lookup_nodes(destination,node)
+        for value in changes:
+            dest.append(value)
 
-    patchers = {
-        PUSH: push,
-        PULL: pull,
-        REMOVE: remove,
-        ADD: add,
-        CHANGE: change
-    }
+    def patch(self,diff_result,destination,copy=True):
+        if copy:
+            dest = deepcopy(destination)
+        else:
+            dest = destination
+        for method, node,changes in diff_result:
+            self.patchers[method](dest,node,changes)
+        return dest
 
-    for action, node, changes in diff_result:
-        patchers[action](node, changes)
+    def _swap_push(self,node , changes):
+        return self.PULL, node, changes
 
-    return destination
+    def _swap_pull(self,node, changes):
+        return self.PUSH, node, changes
 
+    def _swap_add(self,node, changes):
+        return self.REMOVE, node, changes
 
-def swap(diff_result):
-    """
-    Swaps the diff result with the following mapping
+    def _swap_remove(self,node, changes):
+        return self.ADD, node, changes
 
-        pull -> push
-        push -> pull
-        remove -> add
-        add -> remove
-
-    In addition, swaps the changed values for `change` flag.
-
-        >>> swapped = swap([('add', 'a.b.c', ('a', 'b'))])
-        >>> next(swapped)
-        ('remove', 'a.b.c', ('a', 'b'))
-
-        >>> swapped = swap([('change', 'a.b.c', ('a', 'b'))])
-        >>> next(swapped)
-        ('change', 'a.b.c', ('b', 'a'))
-
-    """
-
-    def push(node , changes):
-        return PULL, node, changes
-
-    def pull(node, changes):
-        return PUSH, node, changes
-
-    def add(node, changes):
-        return REMOVE, node, changes
-
-    def remove(node, changes):
-        return ADD, node, changes
-
-    def change(node, changes):
+    def _swap_change(self,node, changes):
         first, second = changes
-        return CHANGE, node, (second, first)
+        return self.CHANGE, node, (second, first)
 
-    swappers = {
-        PUSH: push,
-        PULL: pull,
-        REMOVE: remove,
-        ADD: add,
-        CHANGE: change
-    }
+    def swap(self,diff_result):
+        for method, node, change in diff_result:
+            yield self.swappers[method](node,change)
 
-    for action, node, change in diff_result:
-        yield swappers[action](node, change)
-
-
-def revert(diff_result, destination):
-    """
-    A helper function that calles swap function to revert
-    patched dictionary object.
-
-        >>> first = {'a': 'b'}
-        >>> second = {'a': 'c'}
-        >>> revert(diff(first, second), second)
-        {'a': 'b'}
-
-    """
-    return patch(swap(diff_result), destination)
-
-
-def dot_lookup(source, lookup, parent=False):
-    """
-    A helper function that allows you to reach dictionary
-    items with dot lookup (e.g. document.properties.settings)
-
-        >>> dot_lookup({'a': {'b': 'hello'}}, 'a.b')
-        'hello'
-
-    If parent argument is True, returns the parent node of matched
-    object.
-
-        >>> dot_lookup({'a': {'b': 'hello'}}, 'a.b', parent=True)
-        {'b': 'hello'}
-
-    If node is empy value, returns the whole dictionary object.
-
-        >>> dot_lookup({'a': {'b': 'hello'}}, '')
-        {'a': {'b': 'hello'}}
-
-    """
-    if not lookup:
-        return source
-
-    value = source
-    keys = lookup.split('.')
-
-    if parent:
-        keys = keys[:-1]
-
-    for key in keys:
-        value = value[key]
-    return value
+    def revert(self,diff_result,destination,copy=True):
+        return self.patch(self.swap(diff_result),destination,copy=copy)
